Refactor and explain all steps.
